refactor: Apply changes from suggestions
