+++ app/src/main/java/app/revanced/manager/ui/component/settings/SafeguardBooleanItem.kt
@@ -5,7 +5,10 @@ import androidx.compose.foundation.clickable
 import androidx.compose.material3.Switch
 import androidx.compose.runtime.Composable
 import androidx.compose.runtime.getValue
+import androidx.compose.runtime.mutableStateOf
 import androidx.compose.runtime.rememberCoroutineScope
+import androidx.compose.runtime.saveable.rememberSaveable
+import androidx.compose.runtime.setValue
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.res.stringResource
 import app.revanced.manager.domain.manager.base.Preference
@@ -13,20 +16,39 @@ import kotlinx.coroutines.CoroutineScope
 import kotlinx.coroutines.launch
 
 @Composable
-fun BooleanItem(
+fun SafeguardBooleanItem(
     modifier: Modifier = Modifier,
     preference: Preference<Boolean>,
     coroutineScope: CoroutineScope = rememberCoroutineScope(),
     @StringRes headline: Int,
-    @StringRes description: Int
+    @StringRes description: Int,
+    @StringRes confirmationText: Int = 0
 ) {
     val value by preference.getAsState()
+    var showSafeguardWarning by rememberSaveable {
+        mutableStateOf(false)
+    }
 
-    BooleanItem(
+    if (showSafeguardWarning) {
+        SafeguardConfirmationDialog(
+            onDismiss = { showSafeguardWarning = false },
+            onConfirm = {
+                coroutineScope.launch { preference.update(!value) }
+                showSafeguardWarning = false
+            },
+            body = stringResource(confirmationText)
+        )
+    }
+
+    SafeguardBooleanItem(
         modifier = modifier,
         value = value,
         onValueChange = {
-            coroutineScope.launch { preference.update(it) }
+            if (it != preference.default) {
+                showSafeguardWarning = true
+            } else {
+                coroutineScope.launch { preference.update(it) }
+            }
         },
         headline = headline,
         description = description
@@ -34,7 +56,7 @@ fun BooleanItem(
 }
 
 @Composable
-fun BooleanItem(
+fun SafeguardBooleanItem(
     modifier: Modifier = Modifier,
     value: Boolean,
     onValueChange: (Boolean) -> Unit,
