+++ app/src/main/java/app/revanced/manager/ui/screen/settings/AdvancedSettingsScreen.kt
@@ -39,6 +39,7 @@ import app.revanced.manager.ui.component.ColumnWithScrollbar
 import app.revanced.manager.ui.component.GroupHeader
 import app.revanced.manager.ui.component.settings.BooleanItem
 import app.revanced.manager.ui.component.settings.IntegerItem
+import app.revanced.manager.ui.component.settings.SafeguardBooleanItem
 import app.revanced.manager.ui.component.settings.SettingsListItem
 import app.revanced.manager.ui.viewmodel.AdvancedSettingsViewModel
 import org.koin.androidx.compose.koinViewModel
@@ -121,36 +122,28 @@ fun AdvancedSettingsScreen(
             )
 
             GroupHeader(stringResource(R.string.safeguards))
-            BooleanItem(
-                isSafeguard = true,
-                suggestedValue = false,
+            SafeguardBooleanItem(
                 preference = vm.prefs.disablePatchVersionCompatCheck,
                 coroutineScope = vm.viewModelScope,
                 headline = R.string.patch_compat_check,
                 description = R.string.patch_compat_check_description,
                 confirmationText = R.string.patch_compat_check_confirmation
             )
-            BooleanItem(
-                isSafeguard = true,
-                suggestedValue = false,
+            SafeguardBooleanItem(
                 preference = vm.prefs.disableUniversalPatchWarning,
                 coroutineScope = vm.viewModelScope,
                 headline = R.string.universal_patches_safeguard,
                 description = R.string.universal_patches_safeguard_description,
                 confirmationText = R.string.universal_patches_safeguard_confirmation
             )
-            BooleanItem(
-                isSafeguard = true,
-                suggestedValue = true,
+            SafeguardBooleanItem(
                 preference = vm.prefs.suggestedVersionSafeguard,
                 coroutineScope = vm.viewModelScope,
                 headline = R.string.suggested_version_safeguard,
                 description = R.string.suggested_version_safeguard_description,
                 confirmationText = R.string.suggested_version_safeguard_confirmation
             )
-            BooleanItem(
-                isSafeguard = true,
-                suggestedValue = false,
+            SafeguardBooleanItem(
                 preference = vm.prefs.disableSelectionWarning,
                 coroutineScope = vm.viewModelScope,
                 headline = R.string.patch_selection_safeguard,
